Refactor calendar grid keyboard navigation in separate hook
